rename package disc to discord
